Добавлен класс Loader
