+++ js/application.js
@@ -4,7 +4,7 @@ import RulesView from './templates/rules/rules-view';
 import GameScreen from './templates/game/game-screen';
 import GameModel from './templates/game/game-model';
 import StatisticsView from './templates/statistics/statistics-view';
-import {loadData, loadImages, saveStats, loadResults} from './utils/loader';
+import Loader from './loader/loader';
 
 const main = document.querySelector(`.central`);
 const changeView = (element) => {
@@ -18,12 +18,12 @@ class Application {
   static start() {
     const intro = new IntroView();
     changeView(intro.element);
-    loadData()
+    Loader.loadQuestions()
         .then((data) => {
           gameData = data;
           return data;
         })
-        .then((data) => loadImages(data))
+        .then((data) => Loader.preloadImages(data))
         // анимация исчезновения экрана
         .then(() => intro.remove())
         // появление экрана
@@ -52,8 +52,8 @@ class Application {
   }
 
   static showStats(model, player) {
-    saveStats(model, player)
-        .then(() => loadResults(player))
+    Loader.saveResults(model, player)
+        .then(() => Loader.loadResults(player))
         .then((data) => new StatisticsView(data))
         .then((elem) => changeView(elem.element));
   }
